+++ Wake.Commerce.sln
@@ -5,17 +5,17 @@ VisualStudioVersion = 17.5.33627.172
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.Shared", "src\Wake.Commerce.Shared\Wake.Commerce.Shared.csproj", "{74773402-9457-4EED-A4F3-2E0CE6A23E2B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.Business", "src\Wake.Commerce.Business\Wake.Commerce.Business.csproj", "{8BFD9E81-EDA5-4AE4-9581-538FAA759971}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.Application", "src\Wake.Commerce.Application\Wake.Commerce.Application.csproj", "{8BFD9E81-EDA5-4AE4-9581-538FAA759971}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.Api", "src\Wake.Commerce.Api\Wake.Commerce.Api.csproj", "{A926056D-DA5D-4D1D-AFEE-E48F8A0D3B32}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.UnitTests", "test\Wake.Commerce.UnitTests\Wake.Commerce.UnitTests.csproj", "{238E331F-B412-4F12-92B2-74CB7CF08A95}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.Repository", "src\Wake.Commerce.Repository\Wake.Commerce.Repository.csproj", "{497F682F-A039-4B09-895A-0D6A18D08F5D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.Infrastructure", "src\Wake.Commerce.Infrastructure\Wake.Commerce.Infrastructure.csproj", "{497F682F-A039-4B09-895A-0D6A18D08F5D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wake.Commerce.Domain", "src\Wake.Commerce.Domain\Wake.Commerce.Domain.csproj", "{AC41CAD5-3F58-43B3-B8C8-E9A44868511D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.Domain", "src\Wake.Commerce.Domain\Wake.Commerce.Domain.csproj", "{AC41CAD5-3F58-43B3-B8C8-E9A44868511D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wake.Commerce.IntegrationTests", "test\Wake.Commerce.IntegrationTests\Wake.Commerce.IntegrationTests.csproj", "{61DE757D-301A-4318-9B04-D359F6F950F1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wake.Commerce.IntegrationTests", "test\Wake.Commerce.IntegrationTests\Wake.Commerce.IntegrationTests.csproj", "{61DE757D-301A-4318-9B04-D359F6F950F1}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{864955F1-7112-4F6F-A111-B5712D7FE36D}"
 EndProject
