+++ src/Wake.Commerce.Application/AutoMapper/EntityToDto.cs
@@ -1,0 +1,14 @@
+﻿using AutoMapper;
+using Wake.Commerce.Application.Features.Produtos.Commands.CriarProduto;
+using Wake.Commerce.Domain.Entities;
+
+namespace Wake.Commerce.Application.AutoMapper
+{
+    public class EntityToDto : Profile
+    {
+        public EntityToDto()
+        {
+            CreateMap<CriarProdutoCommand, Produto>();
+        }
+    }
+}
