+++ src/Wake.Commerce.Application/Features/Produtos/Commands/CriarProduto/CriarProdutoCommandVm.cs
@@ -1,0 +1,7 @@
+﻿namespace Wake.Commerce.Application.Features.Produtos.Commands.CriarProduto
+{
+    public class CriarProdutoCommandVm
+    {
+        public int ProdutoId { get; set; }
+    }
+}
