--- src/Wake.Commerce.Business/AutoMapper/EntityToDto.cs
@@ -1,14 +1,0 @@
-﻿using AutoMapper;
-using Wake.Commerce.Business.Features.Produtos.Commands.CriarProduto;
-using Wake.Commerce.Domain.Entities;
-
-namespace Wake.Commerce.Business.AutoMapper
-{
-    public class EntityToDto : Profile
-    {
-        public EntityToDto()
-        {
-            CreateMap<CriarProdutoCommand, Produto>();
-        }
-    }
-}
