--- src/Wake.Commerce.Business/Behaviours/ValidationBehaviour.cs
@@ -1,31 +1,0 @@
-﻿using FluentValidation;
-using MediatR;
-
-namespace Wake.Commerce.Business.Behaviours
-{
-    public class ValidationBehaviour<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
-        where TRequest : IRequest<TResponse>
-    {
-        private readonly IEnumerable<IValidator<TRequest>> _validators;
-
-        public ValidationBehaviour(IEnumerable<IValidator<TRequest>> validators)
-        {
-            _validators = validators;
-        }
-
-        public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
-        {
-            if (_validators.Any())
-            {
-                var context = new ValidationContext<TRequest>(request);
-
-                var validationResults = await Task.WhenAll(_validators.Select(async v => await v.ValidateAsync(context, cancellationToken)).ToList());
-                var failures = validationResults.SelectMany(r => r.Errors).Where(f => f != null).ToList();
-
-                if (failures.Count != 0)
-                    throw new ValidationException(failures);
-            }
-            return await next();
-        }
-    }
-}
