--- src/Wake.Commerce.Business/Features/Produtos/Commands/CriarProduto/CriarProdutoCommandVm.cs
@@ -1,7 +1,0 @@
-﻿namespace Wake.Commerce.Business.Features.Produtos.Commands.CriarProduto
-{
-    public class CriarProdutoCommandVm
-    {
-        public int ProdutoId { get; set; }
-    }
-}
