+++ src/Wake.Commerce.Infrastructure/Interfaces/Repositories/Base/IBaseRepository.cs
@@ -1,0 +1,12 @@
+﻿using Wake.Commerce.Domain.Base;
+
+namespace Wake.Commerce.Infrastructure.Interfaces.Repositories.Base
+{
+    public interface IBaseRepository<T> where T : EntidadeBase
+    {
+        IQueryable<T> GetQuery();
+        Task<IReadOnlyList<T>> GetAllAsync();
+        Task AddAsync(List<T> entity);
+        Task AddAsync(T entity);
+    }
+}
