+++ src/Wake.Commerce.Infrastructure/Interfaces/Repositories/IProdutoRepository.cs
@@ -1,0 +1,9 @@
+﻿using Wake.Commerce.Domain.Entities;
+using Wake.Commerce.Infrastructure.Interfaces.Repositories.Base;
+
+namespace Wake.Commerce.Infrastructure.Interfaces.Repositories
+{
+    public interface IProdutoRepository : IBaseRepository<Produto>
+    {
+    }
+}
