--- src/Wake.Commerce.Repository/Interfaces/Repositories/IProdutoRepository.cs
@@ -1,9 +1,0 @@
-﻿using Wake.Commerce.Domain.Entities;
-using Wake.Commerce.Repository.Interfaces.Repositories.Base;
-
-namespace Wake.Commerce.Repository.Interfaces.Repositories
-{
-    public interface IProdutoRepository : IBaseRepository<Produto>
-    {
-    }
-}
